+++ src/_pytask/node_protocols.py
@@ -14,19 +14,18 @@ if TYPE_CHECKING:
     from _pytask.mark import Mark
 
 
-__all__ = ["MetaNode", "PDelayedNode", "PNode", "PPathNode", "PTask", "PTaskWithPath"]
+__all__ = ["PDelayedNode", "PNode", "PPathNode", "PTask", "PTaskWithPath"]
 
 
 @runtime_checkable
-class MetaNode(Protocol):
-    """Protocol for an intersection between nodes and tasks."""
+class PNode(Protocol):
+    """Protocol for nodes."""
 
     name: str
 
     @property
     def signature(self) -> str:
         """Return the signature of the node."""
-        ...
 
     @abstractmethod
     def state(self) -> str | None:
@@ -36,12 +35,6 @@ class MetaNode(Protocol):
         does not exist, you can also return ``None``.
 
         """
-        ...
-
-
-@runtime_checkable
-class PNode(MetaNode, Protocol):
-    """Protocol for nodes."""
 
     def load(self, is_product: bool) -> Any:
         """Return the value of the node that will be injected into the task.
@@ -55,11 +48,9 @@ class PNode(MetaNode, Protocol):
             user calling :meth:`PNode.load`.
 
         """
-        ...
 
     def save(self, value: Any) -> Any:
         """Save the value that was returned from a task."""
-        ...
 
 
 @runtime_checkable
@@ -74,19 +65,32 @@ class PPathNode(PNode, Protocol):
 
 
 @runtime_checkable
-class PTask(MetaNode, Protocol):
+class PTask(Protocol):
     """Protocol for nodes."""
 
+    name: str
     depends_on: dict[str, PyTree[PNode]]
     produces: dict[str, PyTree[PNode]]
+    function: Callable[..., Any]
     markers: list[Mark]
     report_sections: list[tuple[str, str, str]]
     attributes: dict[Any, Any]
-    function: Callable[..., Any]
+
+    @property
+    def signature(self) -> str:
+        """Return the signature of the node."""
+
+    @abstractmethod
+    def state(self) -> str | None:
+        """Return the state of the node.
+
+        The state can be something like a hash or a last modified timestamp. If the node
+        does not exist, you can also return ``None``.
+
+        """
 
     def execute(self, **kwargs: Any) -> Any:
         """Return the value of the node that will be injected into the task."""
-        ...
 
 
 @runtime_checkable